Add first version of the driver
The driver uses pci to communicate with the guest vmm. As we don't
support dynamic create/destroy of the VMs, the driver comes up with a
vmm pool that is populated as the dataports are probed. This way we can
hide the limitation from the user space as well as the driver core.

Currently destroying VMs, destroying vpci devices, ioeventfds and irqfds
are not supported.

Signed-off-by: Markku Ahvenjärvi <markkux@ssrc.tii.ae>