rpc: use rpc queues for ioreqs
IO requests are initiated by driver (the guest), while most of the other
communication are initiated by the device. Hence it makes sense to
separate communication to own set of queues.

The IO requests use request-response semantics that fit well for lending the
message buffer. As we use fixed sized messages, the response can simply reuse
the request message buffer.

Signed-off-by: Markku Ahvenjärvi <mankku@gmail.com>